// src/components/utils/enum.ts
/* eslint-disable no-unused-vars */
export enum FrontEndEnums {
    Home = '/',
    About = '/about',
    Service = '/services',
    Support = '/support',
    Team = '/team',
    JoinCommunity = '/joincommunity',
    Teamview = '/Teamview',
    Signin = '/signin',
    Welcome = '/welcome',
    Addservice = '/addservice',
    EmployeeHome = '/employeehome',
    FreelanceHome = '/freelancehome',
    Freelances = '/freelances',
    Addemployee ='/addemployee',
    Addfreelance ='/addfreelance',
    Newservice ='/newservice',
    Employeeall= '/employeeall',
    Teamcat= '/temcategory',
    Serviceadd = '/serviceadd',
    Ourfreelance = '/ourteam'

}
export enum backEndPoints {
    LOGIN = 'login',
    USERCREATION = 'user-creation',
    SERVICE = 'service',
    COMMUNITY = 'team',
    COMMUNITYCAT = 'team/category',
    EMPLOYEE = 'employer',
    INTOUCH = 'intouch',
    IMAGE = 'image'
  }
export enum UserType {
    ADMIN = 'Admin'
}

export enum freelanceToken {
    userToken = 'token',
    userId = 'id',
    userEmail = 'email',
    userName = 'fullname',
    accountType = 'accountType'

}